requirement.txt
# Core Data Processing
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
urllib3>=1.26.0

# Web Scraping & APIs
google-api-python-client>=2.0.0
snscrape>=0.6.0
feedparser>=6.0.0
playwright>=1.30.0
beautifulsoup4>=4.11.0

# NLP & ML Libraries
transformers>=4.21.0
torch>=1.12.0
spacy>=3.4.0
scikit-learn>=1.1.0
nltk>=3.7
sumy>=0.11.0

# Text Processing
regex>=2022.7.0
html5lib>=1.1
lxml>=4.9.0

# Visualization & UI
streamlit>=1.25.0
plotly>=5.10.0
matplotlib>=3.5.0
seaborn>=0.11.0
wordcloud>=1.9.0

# Data Storage & Serialization
joblib>=1.2.0
pickle-mixin>=1.0.2

# Utilities
python-dateutil>=2.8.0
pytz>=2022.1
certifi>=2022.6.0
charset-normalizer>=2.1.0

# Optional Dependencies (for advanced features)
# fastapi[all]>=0.85.0  # For API endpoints
# python-jose>=3.3.0    # For JWT tokens
# tweepy>=4.12.0        # For Twitter API (alternative)
# instaloader>=4.9.0    # For Instagram scraping
# kaggle>=1.5.0         # For Kaggle datasets
# reportlab>=3.6.0      # For PDF generation

# Development Dependencies
# pytest>=7.1.0
# black>=22.6.0
# flake8>=5.0.0
# mypy>=0.971